refactor: 가격 계산 함수 custom hook 으로 분리
